requirements.txt
# Core
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0

# OpenAI / Ollama
langchain-openai>=0.0.2
langchain-ollama>=0.0.1

# Vector Store
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Document Loaders
pypdf>=3.0.0
python-docx>=1.0.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 数据库和认证
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
